+++ vignettes/WASP.R
@@ -1,21 +1,4 @@
----
-title: "WASP： An R package for Wavelet System Prediction"
-author: 'Ze Jiang, Ashish Sharma, and Fiona Johnson'
-date: "`r format(Sys.time(), '%H:%M:%S %d %B, %Y')`"
-output: 
-#  rmarkdown::html_vignette: 
-  bookdown::html_vignette2: 
-    toc: true
-    number_sections: true
-vignette: >
-  %\VignetteIndexEntry{Wavelet System Prediction}
-  %\VignetteEncoding{UTF-8}
-  %\VignetteEngine{knitr::rmarkdown}
-editor_options: 
-  chunk_output_type: console
----
-
-```{r, include = FALSE}
+## ---- include = FALSE---------------------------------------------------------
 knitr::opts_chunk$set(
   collapse = TRUE,
   comment = "#>", warning = FALSE, 
@@ -23,12 +6,12 @@ knitr::opts_chunk$set(
   out.width = "85%",
   fig.align = "center", fig.pos = "h!"
 )
+options(rmarkdown.html_vignette.check_title = FALSE)
 library(rmarkdown)
 library(knitr)
 library(kableExtra)
-```
 
-```{r packages}
+## ----packages-----------------------------------------------------------------
 library(WASP)
 library(ggplot2)
 
@@ -37,11 +20,8 @@ library(synthesis)
 library(waveslim)
 library(cowplot)
 library(gridGraphics)
-```
 
-# DWT, MODWT and AT basic propertites
-
-```{r wavelet-transforms}
+## ----wavelet-transforms-------------------------------------------------------
 # data generation
 x <- arima.sim(list(order = c(1, 0, 0), ar = 0.6), n = 512)
 # x <- as.numeric(scale(data.gen.Rossler(time = seq(0, 50, length.out = 512))$x, scale=F))
@@ -103,11 +83,8 @@ for (wf in c("haar", "d4", "d8", "d16")) {
     message(paste0("AT and MODWT is equivalent using the", wf, "!"))
   }
 }
-```
 
-## Summary of various properties for the three DWT methods
-
-```{r tab1}
+## ----tab1---------------------------------------------------------------------
 tab1 <- data.frame(
   col1 = c("DWT-MRA", "MODWT", "AT"),
   col2 = c("$\\checkmark$", "", "$\\checkmark$"),
@@ -123,11 +100,8 @@ kable(tab1, caption = "Summary of various properties for the three DWT methods",
   column_spec(1, width = "6em") %>%
   column_spec(2:5, width = "7em") %>%
   footnote(general = "When Haar wavelet filter is used, MODWT and AT are equivalent and both of them preserves additive and variance decomposition.", footnote_as_chunk = T)
-```
 
-## Illustration of three types of DWT methods 
-
-```{r wavelet-decomposition, fig.show='hide'}
+## ----wavelet-decomposition, fig.show='hide'-----------------------------------
 p.list <- NULL
 wf.opts <- c("d16", "haar")
 for (k in seq_along(wf.opts)) {
@@ -179,33 +153,24 @@ for (k in seq_along(wf.opts)) {
 
   p.list[[k]] <- list(p1, p2, p3)
 }
-```
 
-###  Daubechies 16 wavelet
-
-```{r figa, fig.cap='Illustration of three types of DWT methods', fig.width=9, fig.height=7}
+## ----figa, fig.cap='Illustration of three types of DWT methods', fig.width=9, fig.height=7----
 #----------------------------------------------------------------------------
 # plot and save
 cowplot::plot_grid(
   plotlist = p.list[[1]], ncol = 3, labels = c("(a)", "(b)", "(c)"),
   label_size = 12
 )
-```
 
-### Haar wavelet filter
-
-```{r figb, fig.cap='Illustration of three types of DWT methods', fig.width=9, fig.height=7}
+## ----figb, fig.cap='Illustration of three types of DWT methods', fig.width=9, fig.height=7----
 #----------------------------------------------------------------------------
 # plot and save
 cowplot::plot_grid(
   plotlist = p.list[[2]], ncol = 3, labels = c("(a)", "(b)", "(c)"),
   label_size = 12
 )
-```
 
-## Wavelet transform: decompostion level 
-
-```{r wt-decomposition-level}
+## ----wt-decomposition-level---------------------------------------------------
 sample <- seq(100, by = 200, length.out = 5)
 v <- 2 # vanishing moment
 tmp <- NULL
@@ -230,13 +195,8 @@ kable(tab,
 ) %>%
   kable_styling("striped", position = "center", full_width = FALSE) # %>%
 # collapse_rows(columns = 1:2, valign = "middle")
-```
 
-# Variance transformation
-
-## Optimal preditive accuracy (RMSE)
-
-```{r optimal-variance-transformation, warning=TRUE}
+## ----optimal-variance-transformation, warning=TRUE----------------------------
 if (TRUE) {
   ### Synthetic example
   # data generation
@@ -337,18 +297,14 @@ kable(tab.list[[1]], caption = "Optimal RMSE using DWT-based VT",
       booktabs = T, align = "c", digits = 3) %>%
 kable_styling("striped", position = "center", full_width = FALSE)  %>%
 collapse_rows(columns = 1, valign = "middle")
-```
 
-```{r optimal-variance-transformation1, warning=TRUE}
+## ----optimal-variance-transformation1, warning=TRUE---------------------------
 kable(tab.list[[2]], caption = "Optimal RMSE using MODWT/AT-based VT",
       booktabs = T, align = "c", digits = 3) %>%
 kable_styling("striped", position = "center", full_width = FALSE)  %>%
 collapse_rows(columns = 1, valign = "middle")
-```
 
-## Transformed predictor variables
-
-```{r variance-transform, fig.keep='last', fig.cap='Orignal and VT predictors. (a): DWT-MRA (b): MODWT/AT', fig.width=9, fig.height=6}
+## ----variance-transform, fig.keep='last', fig.cap='Orignal and VT predictors. (a): DWT-MRA (b): MODWT/AT', fig.width=9, fig.height=6----
 #-------------------------------------------------------------------
 if (TRUE) {
   set.seed(2020)
@@ -461,11 +417,8 @@ for (mode in mode.opts) {
 # plot and save
 fig <- cowplot::plot_grid(plotlist = p.list, nrow = 1, labels = c("(a)", "(b)", "(c)"))
 fig
-```
 
-# Stepwise variance transfromation
-
-```{r svt, fig.keep='last', fig.cap='Orignal and SVT predictors. (a): DWT-MRA (b): MODWT/AT', fig.width=9, fig.height=6}
+## ----svt, fig.keep='last', fig.cap='Orignal and SVT predictors. (a): DWT-MRA (b): MODWT/AT', fig.width=9, fig.height=6----
 #-------------------------------------------------------------------
 ### Real-world example
 data("obs.mon")
@@ -576,257 +529,254 @@ kable(tab1, caption = "Comparison of prediction accuracy using Std and SVT", boo
   kable_styling(latex_options = c("HOLD_position"), position = "center", full_width = FALSE)  %>%
   #  add_header_above(c(" " = 1, "DWT-MRA" = 2, "MODWT" = 2, "AT" = 2))
   add_header_above(c(" " = 1, "DWT-MRA" = 2, "MODWT/AT" = 2))
-```
 
-<!-- # Comparison with traditional wavelet-based methods -->
+## ----comp, eval=FALSE, include=FALSE------------------------------------------
+#  #-------------------------------------------------------------------
+#  sample <-  100000
+#  sample.cal <- sample/2
+#  k <- ceiling(sqrt(sample/2))
+#  
+#  s=0.1
+#  #s=c(0.1,0.5,1) # scaling factor for noise level
+#  set.seed(2020)
+#  
+#  ###synthetic example - Rossler
+#  ts.list <- list()
+#  for(i in seq_along(s)){
+#    ts.r <- data.gen.Rossler(a = 0.2, b = 0.2, w = 5.7, start = c(-2, -10, 0.2), time = seq(0, 50, length.out = sample))
+#  
+#    #add noise
+#    ts.r$x <- ts(ts.r$x + rnorm(n = sample, mean=0, sd=s[i]))
+#    ts.r$y <- ts(ts.r$y + rnorm(n = sample, mean=0, sd=s[i]))
+#    ts.r$z <- ts(ts.r$z + rnorm(n = sample, mean=0, sd=s[i]))
+#  
+#    ts.list[[i]]<- ts.r
+#  }
+#  
+#  #-------------------------------------------------------------------
+#  tab3<-NULL
+#  mode.opts <- c("MRA", "MODWT","a trous")[1:2]
+#  for(mode in mode.opts){
+#    ### wavelet method selection
+#    #mode <- switch(3,"MRA", "MODWT","a trous")
+#    cov.opt <- switch(1,"auto","pos","neg")
+#    if(mode=="MRA") method <- switch(1,"dwt","modwt")
+#  
+#    # wavelet family, extension mode and package
+#    wf <- "haar" # wavelet family D8 or db4
+#    pad <-  "zero"
+#    boundary <- "periodic"
+#    if(wf!="haar") v <- as.integer(as.numeric(substr(wf,2,3))/2) else v <- 1
+#  
+#  
+#    ###proposed method----------------------------------------------------------
+#    #--------------------------------------------------
+#    #calibration dataset
+#    data.list <- lapply(ts.list, function(ts) list(x=ts$z[1:sample.cal], dp=cbind(ts$x[1:sample.cal],ts$y[1:sample.cal])))
+#  
+#    n <- sample.cal
+#    J <- ceiling(log(n/(2*v-1))/log(2)) - 1
+#    #if(wf=="haar"&&mode=="MODWT") J = J-1 #since modwt no need a dyadic number size
+#    print(paste0("Calibration: Decomposition Levels J= ",J))
+#  
+#    #variance transfrom
+#    if(mode=="MRA"){
+#      dwt.list<- lapply(data.list, function(x) dwt.vt(x, wf, J, method, pad, boundary, cov.opt))
+#    } else if(mode=="MODWT") {
+#      dwt.list<- lapply(data.list, function(x) modwt.vt(x, wf, J, boundary, cov.opt))
+#    } else {
+#      dwt.list<- lapply(data.list, function(x) at.vt(x, wf, J, boundary, cov.opt))
+#    }
+#  
+#    #--------------------------------------------------
+#    # calibration
+#    df <- NULL;data.RMSE<-NULL;dwt.RMSE<-NULL
+#    sd.cal<-NULL; cor.cal<-NULL
+#    for(i in 1:length(dwt.list)){
+#  
+#      dwt <- dwt.list[[i]]
+#      dp <- dwt$dp; dp.n <- dwt$dp.n; x <- dwt$x
+#  
+#      m1 <- FNN::knn.reg(dp, y=x, k=k)$pred
+#      m2 <- FNN::knn.reg(dp.n, y=x, k=k)$pred
+#  
+#      data.RMSE <-c(data.RMSE, round(sqrt(mean((x-m1)^2)),3))
+#      dwt.RMSE <- c(dwt.RMSE, round(sqrt(mean((x-m2)^2)),3))
+#  
+#      sd.cal <- cbind(sd.cal, as.vector(c(sd(x),sd(m1),sd(m2))))
+#      cor.cal <-cbind(cor.cal, cor(cbind(x,m1,m2))[,1])
+#  
+#      df1 <- data.frame(Group=1, s=s[i], No=1:sample.cal,Pred=m1, Obs=x)
+#      df2 <- data.frame(Group=2, s=s[i], No=1:sample.cal,Pred=m2, Obs=x)
+#  
+#      df <- rbind(df, rbind(df1,df2))
+#  
+#    }
+#    #summary(df)
+#    #print(rbind(data.RMSE,dwt.RMSE))
+#  
+#    t1 <- rbind(data.RMSE,dwt.RMSE)
+#    sd.cal;cor.cal
+#  
+#    #--------------------------------------------------
+#    #validataion dataset
+#    data.list.val <- lapply(ts.list, function(ts) list(x=ts$z[(sample.cal+1):sample], dp=cbind(ts$x[(sample.cal+1):sample], ts$y[(sample.cal+1):sample])))
+#  
+#    sample.val <- sample-sample.cal
+#    n <- sample.val
+#    J <- ceiling(log(n/(2*v-1))/log(2)) - 1
+#    #if(wf=="haar"&&mode=="MODWT") J = J-1 #since modwt no need a dyadic number size
+#    print(paste0("Validation: Decomposition Levels J= ",J))
+#  
+#    #--------------------------------------------------
+#    #variance transform
+#    if(mode=="MRA"){
+#      dwt.list.val<- lapply(1:length(data.list.val), function(i) dwt.vt.val(data.list.val[[i]], J, dwt.list[[i]]))
+#    } else if(mode=="MODWT"){
+#      dwt.list.val<- lapply(1:length(data.list.val), function(i) modwt.vt.val(data.list.val[[i]], J, dwt.list[[i]]))
+#    } else {
+#      dwt.list.val<- lapply(1:length(data.list.val), function(i) at.vt.val(data.list.val[[i]], J, dwt.list[[i]]))
+#    }
+#  
+#    #--------------------------------------------------
+#    # validation
+#    df.val <- NULL;data.RMSE <-NULL;dwt.RMSE<-NULL
+#    sd.val<-NULL; cor.val<-NULL
+#    for(i in 1:length(dwt.list.val)){
+#  
+#      dwt <- dwt.list[[i]]
+#      dp <- dwt$dp; dp.n <- dwt$dp.n; x.train <- dwt$x
+#  
+#      dwt <- dwt.list.val[[i]]
+#      dp.v <- dwt$dp; dp.n.v <- dwt$dp.n; x <- dwt$x
+#  
+#      m1 <- FNN::knn.reg(train=dp, test=dp.v, y=x.train, k=k)$pred
+#      m2 <- FNN::knn.reg(train=dp.n, test=dp.n.v, y=x.train, k=k)$pred
+#  
+#      data.RMSE <-c(data.RMSE, round(sqrt(mean((m1-x)^2)),3))
+#      dwt.RMSE <- c(dwt.RMSE, round(sqrt(mean((m2-x)^2)),3))
+#  
+#      sd.val <- cbind(sd.val, as.vector(c(sd(x),sd(m1),sd(m2))))
+#      cor.val <- cbind(cor.val, cor(cbind(x,m1,m2))[,1])
+#  
+#      df1 <- data.frame(Group=1, s=s[i], No=1:sample.val,Pred=m1, Obs=x)
+#      df2 <- data.frame(Group=2, s=s[i], No=1:sample.val,Pred=m2, Obs=x)
+#  
+#      df.val <- rbind(df.val, rbind(df1,df2))
+#  
+#    }
+#  
+#    #summary(df.val)
+#    #print(rbind(data.RMSE,dwt.RMSE))
+#  
+#    t2 <- rbind(data.RMSE,dwt.RMSE)
+#    sd.val;cor.val
+#  
+#    ###standard method----------------------------------------------------------
+#    # form new response and predictors dataset - calibration
+#    data.list <- list()
+#    for(i in 1:length(ts.list)){
+#      #i <- 1
+#      x <- ts.list[[i]]$x[1:sample.cal]
+#      y <- ts.list[[i]]$y[1:sample.cal]
+#      z <- ts.list[[i]]$z[1:sample.cal]
+#  
+#      xx <- padding(x, pad); yy <- padding(y, pad)
+#      n <- length(x)
+#  
+#      J <- floor(log10(n)) # (Nourani et al., 2008)
+#      print(paste0("Direct wavelet approach: Decomposition Levels J= ",J))
+#  
+#      if(mode=="MRA"){
+#      mra.x <- matrix(unlist(lapply(mra(xx,wf,J,method,boundary), function(z) z[1:n])), ncol=J+1, byrow=FALSE)
+#      mra.y <- matrix(unlist(lapply(mra(yy,wf,J,method,boundary), function(z) z[1:n])), ncol=J+1, byrow=FALSE)
+#      } else if(mode=="MODWT"){
+#      mra.x <- matrix(unlist(lapply(modwt(x,wf,J,boundary), function(z) z)), ncol=J+1, byrow=FALSE)
+#      mra.y <- matrix(unlist(lapply(modwt(y,wf,J,boundary), function(z) z)), ncol=J+1, byrow=FALSE)
+#      } else {
+#      mra.x <- matrix(unlist(at.wd(x,wf,J,boundary)), ncol=J+1)
+#      mra.y <- matrix(unlist(at.wd(y,wf,J,boundary)), ncol=J+1)
+#      }
+#  
+#      data.list[[i]] <- list(x=as.numeric(z), dp=cbind(mra.x, mra.y))
+#    }
+#  
+#    #----------------------------------------------------
+#    #calibration
+#    df <- NULL;data.RMSE<-NULL
+#    for(i in 1:length(data.list)){
+#      dwt <- data.list[[i]]
+#      x <- dwt$x; dp <- dwt$dp
+#  
+#      m <- FNN::knn.reg(train=dp, y=x, k=k)$pred
+#      data.RMSE <-c(data.RMSE, round(sqrt(mean((m-x)^2)),3))
+#      df <- data.frame(Group=1, s=s[i], No=1:sample.cal,Pred=m, Obs=x)
+#    }
+#  
+#    #----------------------------------------------------
+#    # form new response and predictors dataset -  validation
+#    data.list.val <- list()
+#    for(i in 1:length(ts.list)){
+#      #i <- 1
+#      x <- ts.list[[i]]$x[(sample.cal+1):sample]
+#      y <- ts.list[[i]]$y[(sample.cal+1):sample]
+#      z <- ts.list[[i]]$z[(sample.cal+1):sample]
+#  
+#      xx <- padding(x, pad); yy <- padding(y, pad)
+#      n <- length(x)
+#  
+#      J <- floor(log10(n)) # (Nourani et al., 2008)
+#  
+#      if(mode=="MRA"){
+#      mra.x <- matrix(unlist(lapply(mra(xx,wf,J,method,boundary), function(z) z[1:n])), ncol=J+1, byrow=FALSE)
+#      mra.y <- matrix(unlist(lapply(mra(yy,wf,J,method,boundary), function(z) z[1:n])), ncol=J+1, byrow=FALSE)
+#      } else if(mode=="MODWT"){
+#      mra.x <- matrix(unlist(lapply(modwt(x,wf,J,boundary), function(z) z)), ncol=J+1, byrow=FALSE)
+#      mra.y <- matrix(unlist(lapply(modwt(y,wf,J,boundary), function(z) z)), ncol=J+1, byrow=FALSE)
+#      } else {
+#      mra.x <- matrix(unlist(at.wd(x,wf,J,boundary)), ncol=J+1)
+#      mra.y <- matrix(unlist(at.wd(y,wf,J,boundary)), ncol=J+1)
+#      }
+#  
+#      data.list.val[[i]] <- list(x=as.numeric(z), dp=cbind(mra.x, mra.y))
+#    }
+#  
+#    #----------------------------------------------------
+#    #validation
+#    sample.val <- sample-sample.cal
+#    df.val <- NULL;dwt.RMSE<-NULL
+#    for(i in 1:length(data.list.val)){
+#      dwt <- data.list[[i]]
+#      x.train <- dwt$x; dp <- dwt$dp
+#  
+#      dwt <- data.list.val[[i]]
+#      x <- dwt$x; dp.v <- dwt$dp
+#  
+#      m <- FNN::knn.reg(train=dp, test=dp.v, y=x.train, k=k)$pred
+#      dwt.RMSE <-c(dwt.RMSE, round(sqrt(mean((m-x)^2)),3))
+#      df.val <- data.frame(Group=1, s=s[i], No=1:sample.val,Pred=m, Obs=x)
+#    }
+#  
+#    t3 <- rbind(data.RMSE,dwt.RMSE)
+#  
+#    #----------------------------------------------------
+#    #comparison
+#    df.RMSE <- rbind(rbind(t1,t2),t3)
+#    rownames(df.RMSE) <- NULL
+#    df.RMSE.n <- data.frame(Method=mode,
+#      Group=c("Calibration", "Calibration", "Validation", "Validation",
+#              "Calibration", "Validation"),
+#      Model = c("Original", "VT", "Original", "VT", "Wavelet-decomposed components",
+#                "Wavelet-decomposed components"),df.RMSE)%>%
+#      tidyr::gather(S,Value,4:(3+length(s)))%>% tidyr::spread(Group, Value)
+#  
+#    tab3 <- rbind(tab3,df.RMSE.n[order(df.RMSE.n$S),])
+#  
+#  }
+#  
+#  #----------------------------------------------------
+#  kable(tab3[,-3], caption= "Comparison of three methods using original predictor,
+#        wavelet-decomposed components, and variance-transformed predictor", booktabs = T)%>%
+#  kable_styling("striped", position = "center", full_width = FALSE) %>%
+#  collapse_rows(columns = 1, valign = "middle")
 
-```{r comp, eval=FALSE, include=FALSE}
-#-------------------------------------------------------------------
-sample <-  100000
-sample.cal <- sample/2
-k <- ceiling(sqrt(sample/2))
-
-s=0.1
-#s=c(0.1,0.5,1) # scaling factor for noise level
-set.seed(2020)
-
-###synthetic example - Rossler
-ts.list <- list()
-for(i in seq_along(s)){
-  ts.r <- data.gen.Rossler(a = 0.2, b = 0.2, w = 5.7, start = c(-2, -10, 0.2), time = seq(0, 50, length.out = sample))
-  
-  #add noise
-  ts.r$x <- ts(ts.r$x + rnorm(n = sample, mean=0, sd=s[i]))
-  ts.r$y <- ts(ts.r$y + rnorm(n = sample, mean=0, sd=s[i]))
-  ts.r$z <- ts(ts.r$z + rnorm(n = sample, mean=0, sd=s[i]))
-  
-  ts.list[[i]]<- ts.r
-}
-
-#-------------------------------------------------------------------
-tab3<-NULL
-mode.opts <- c("MRA", "MODWT","a trous")[1:2]
-for(mode in mode.opts){
-  ### wavelet method selection
-  #mode <- switch(3,"MRA", "MODWT","a trous")
-  cov.opt <- switch(1,"auto","pos","neg")
-  if(mode=="MRA") method <- switch(1,"dwt","modwt")
-
-  # wavelet family, extension mode and package
-  wf <- "haar" # wavelet family D8 or db4
-  pad <-  "zero"
-  boundary <- "periodic"
-  if(wf!="haar") v <- as.integer(as.numeric(substr(wf,2,3))/2) else v <- 1
-  
-
-  ###proposed method----------------------------------------------------------
-  #--------------------------------------------------
-  #calibration dataset
-  data.list <- lapply(ts.list, function(ts) list(x=ts$z[1:sample.cal], dp=cbind(ts$x[1:sample.cal],ts$y[1:sample.cal])))
-  
-  n <- sample.cal
-  J <- ceiling(log(n/(2*v-1))/log(2)) - 1
-  #if(wf=="haar"&&mode=="MODWT") J = J-1 #since modwt no need a dyadic number size
-  print(paste0("Calibration: Decomposition Levels J= ",J))
-  
-  #variance transfrom
-  if(mode=="MRA"){
-    dwt.list<- lapply(data.list, function(x) dwt.vt(x, wf, J, method, pad, boundary, cov.opt))
-  } else if(mode=="MODWT") {
-    dwt.list<- lapply(data.list, function(x) modwt.vt(x, wf, J, boundary, cov.opt))
-  } else {
-    dwt.list<- lapply(data.list, function(x) at.vt(x, wf, J, boundary, cov.opt))
-  }
-  
-  #--------------------------------------------------
-  # calibration
-  df <- NULL;data.RMSE<-NULL;dwt.RMSE<-NULL
-  sd.cal<-NULL; cor.cal<-NULL
-  for(i in 1:length(dwt.list)){
-  
-    dwt <- dwt.list[[i]]
-    dp <- dwt$dp; dp.n <- dwt$dp.n; x <- dwt$x
-  
-    m1 <- FNN::knn.reg(dp, y=x, k=k)$pred
-    m2 <- FNN::knn.reg(dp.n, y=x, k=k)$pred
-  
-    data.RMSE <-c(data.RMSE, round(sqrt(mean((x-m1)^2)),3))
-    dwt.RMSE <- c(dwt.RMSE, round(sqrt(mean((x-m2)^2)),3))
-  
-    sd.cal <- cbind(sd.cal, as.vector(c(sd(x),sd(m1),sd(m2))))
-    cor.cal <-cbind(cor.cal, cor(cbind(x,m1,m2))[,1])
-  
-    df1 <- data.frame(Group=1, s=s[i], No=1:sample.cal,Pred=m1, Obs=x)
-    df2 <- data.frame(Group=2, s=s[i], No=1:sample.cal,Pred=m2, Obs=x)
-  
-    df <- rbind(df, rbind(df1,df2))
-  
-  }
-  #summary(df)
-  #print(rbind(data.RMSE,dwt.RMSE))
-  
-  t1 <- rbind(data.RMSE,dwt.RMSE)
-  sd.cal;cor.cal
-  
-  #--------------------------------------------------
-  #validataion dataset
-  data.list.val <- lapply(ts.list, function(ts) list(x=ts$z[(sample.cal+1):sample], dp=cbind(ts$x[(sample.cal+1):sample], ts$y[(sample.cal+1):sample])))
-  
-  sample.val <- sample-sample.cal
-  n <- sample.val
-  J <- ceiling(log(n/(2*v-1))/log(2)) - 1
-  #if(wf=="haar"&&mode=="MODWT") J = J-1 #since modwt no need a dyadic number size
-  print(paste0("Validation: Decomposition Levels J= ",J))
-  
-  #--------------------------------------------------
-  #variance transform
-  if(mode=="MRA"){
-    dwt.list.val<- lapply(1:length(data.list.val), function(i) dwt.vt.val(data.list.val[[i]], J, dwt.list[[i]]))
-  } else if(mode=="MODWT"){
-    dwt.list.val<- lapply(1:length(data.list.val), function(i) modwt.vt.val(data.list.val[[i]], J, dwt.list[[i]]))
-  } else {
-    dwt.list.val<- lapply(1:length(data.list.val), function(i) at.vt.val(data.list.val[[i]], J, dwt.list[[i]]))
-  }
-  
-  #--------------------------------------------------
-  # validation
-  df.val <- NULL;data.RMSE <-NULL;dwt.RMSE<-NULL
-  sd.val<-NULL; cor.val<-NULL
-  for(i in 1:length(dwt.list.val)){
-  
-    dwt <- dwt.list[[i]]
-    dp <- dwt$dp; dp.n <- dwt$dp.n; x.train <- dwt$x
-  
-    dwt <- dwt.list.val[[i]]
-    dp.v <- dwt$dp; dp.n.v <- dwt$dp.n; x <- dwt$x
-  
-    m1 <- FNN::knn.reg(train=dp, test=dp.v, y=x.train, k=k)$pred
-    m2 <- FNN::knn.reg(train=dp.n, test=dp.n.v, y=x.train, k=k)$pred
-  
-    data.RMSE <-c(data.RMSE, round(sqrt(mean((m1-x)^2)),3))
-    dwt.RMSE <- c(dwt.RMSE, round(sqrt(mean((m2-x)^2)),3))
-  
-    sd.val <- cbind(sd.val, as.vector(c(sd(x),sd(m1),sd(m2))))
-    cor.val <- cbind(cor.val, cor(cbind(x,m1,m2))[,1])
-  
-    df1 <- data.frame(Group=1, s=s[i], No=1:sample.val,Pred=m1, Obs=x)
-    df2 <- data.frame(Group=2, s=s[i], No=1:sample.val,Pred=m2, Obs=x)
-  
-    df.val <- rbind(df.val, rbind(df1,df2))
-  
-  }
-  
-  #summary(df.val)
-  #print(rbind(data.RMSE,dwt.RMSE))
-  
-  t2 <- rbind(data.RMSE,dwt.RMSE)
-  sd.val;cor.val
-  
-  ###standard method----------------------------------------------------------
-  # form new response and predictors dataset - calibration
-  data.list <- list()
-  for(i in 1:length(ts.list)){
-    #i <- 1
-    x <- ts.list[[i]]$x[1:sample.cal]
-    y <- ts.list[[i]]$y[1:sample.cal]
-    z <- ts.list[[i]]$z[1:sample.cal]
-    
-    xx <- padding(x, pad); yy <- padding(y, pad)
-    n <- length(x)
-    
-    J <- floor(log10(n)) # (Nourani et al., 2008)
-    print(paste0("Direct wavelet approach: Decomposition Levels J= ",J))
-
-    if(mode=="MRA"){
-    mra.x <- matrix(unlist(lapply(mra(xx,wf,J,method,boundary), function(z) z[1:n])), ncol=J+1, byrow=FALSE)
-    mra.y <- matrix(unlist(lapply(mra(yy,wf,J,method,boundary), function(z) z[1:n])), ncol=J+1, byrow=FALSE)
-    } else if(mode=="MODWT"){
-    mra.x <- matrix(unlist(lapply(modwt(x,wf,J,boundary), function(z) z)), ncol=J+1, byrow=FALSE)
-    mra.y <- matrix(unlist(lapply(modwt(y,wf,J,boundary), function(z) z)), ncol=J+1, byrow=FALSE)      
-    } else {
-    mra.x <- matrix(unlist(at.wd(x,wf,J,boundary)), ncol=J+1)
-    mra.y <- matrix(unlist(at.wd(y,wf,J,boundary)), ncol=J+1)
-    }
-    
-    data.list[[i]] <- list(x=as.numeric(z), dp=cbind(mra.x, mra.y))
-  }
-  
-  #----------------------------------------------------
-  #calibration
-  df <- NULL;data.RMSE<-NULL
-  for(i in 1:length(data.list)){
-    dwt <- data.list[[i]]
-    x <- dwt$x; dp <- dwt$dp
-    
-    m <- FNN::knn.reg(train=dp, y=x, k=k)$pred
-    data.RMSE <-c(data.RMSE, round(sqrt(mean((m-x)^2)),3))
-    df <- data.frame(Group=1, s=s[i], No=1:sample.cal,Pred=m, Obs=x)
-  }
-  
-  #----------------------------------------------------
-  # form new response and predictors dataset -  validation
-  data.list.val <- list()
-  for(i in 1:length(ts.list)){
-    #i <- 1
-    x <- ts.list[[i]]$x[(sample.cal+1):sample]
-    y <- ts.list[[i]]$y[(sample.cal+1):sample]
-    z <- ts.list[[i]]$z[(sample.cal+1):sample]
-    
-    xx <- padding(x, pad); yy <- padding(y, pad)
-    n <- length(x)
-    
-    J <- floor(log10(n)) # (Nourani et al., 2008)
-    
-    if(mode=="MRA"){
-    mra.x <- matrix(unlist(lapply(mra(xx,wf,J,method,boundary), function(z) z[1:n])), ncol=J+1, byrow=FALSE)
-    mra.y <- matrix(unlist(lapply(mra(yy,wf,J,method,boundary), function(z) z[1:n])), ncol=J+1, byrow=FALSE)
-    } else if(mode=="MODWT"){
-    mra.x <- matrix(unlist(lapply(modwt(x,wf,J,boundary), function(z) z)), ncol=J+1, byrow=FALSE)
-    mra.y <- matrix(unlist(lapply(modwt(y,wf,J,boundary), function(z) z)), ncol=J+1, byrow=FALSE)      
-    } else {
-    mra.x <- matrix(unlist(at.wd(x,wf,J,boundary)), ncol=J+1)
-    mra.y <- matrix(unlist(at.wd(y,wf,J,boundary)), ncol=J+1)
-    }
-
-    data.list.val[[i]] <- list(x=as.numeric(z), dp=cbind(mra.x, mra.y))
-  }
-  
-  #----------------------------------------------------
-  #validation
-  sample.val <- sample-sample.cal
-  df.val <- NULL;dwt.RMSE<-NULL
-  for(i in 1:length(data.list.val)){
-    dwt <- data.list[[i]]
-    x.train <- dwt$x; dp <- dwt$dp
-    
-    dwt <- data.list.val[[i]]
-    x <- dwt$x; dp.v <- dwt$dp
-    
-    m <- FNN::knn.reg(train=dp, test=dp.v, y=x.train, k=k)$pred
-    dwt.RMSE <-c(dwt.RMSE, round(sqrt(mean((m-x)^2)),3))
-    df.val <- data.frame(Group=1, s=s[i], No=1:sample.val,Pred=m, Obs=x)
-  }
-  
-  t3 <- rbind(data.RMSE,dwt.RMSE)
-  
-  #----------------------------------------------------
-  #comparison
-  df.RMSE <- rbind(rbind(t1,t2),t3)
-  rownames(df.RMSE) <- NULL
-  df.RMSE.n <- data.frame(Method=mode,
-    Group=c("Calibration", "Calibration", "Validation", "Validation", 
-            "Calibration", "Validation"),
-    Model = c("Original", "VT", "Original", "VT", "Wavelet-decomposed components", 
-              "Wavelet-decomposed components"),df.RMSE)%>%
-    tidyr::gather(S,Value,4:(3+length(s)))%>% tidyr::spread(Group, Value)
-  
-  tab3 <- rbind(tab3,df.RMSE.n[order(df.RMSE.n$S),])
-
-}
-
-#----------------------------------------------------
-kable(tab3[,-3], caption= "Comparison of three methods using original predictor, 
-      wavelet-decomposed components, and variance-transformed predictor", booktabs = T)%>%
-kable_styling("striped", position = "center", full_width = FALSE) %>% 
-collapse_rows(columns = 1, valign = "middle")
-```