+++ NEWS.md
@@ -1,7 +1,10 @@
+# WASP (development version)
+* Include Authors@R
+* Update README.md and WASP.Rmd
+
 # WASP 1.4.1
 * Update vignettes
 
 # WASP 1.4.0
-
 * Added a `NEWS.md` file to track changes to the package.
 * Initial release on CRAN